+++ actionpack/test/controller/caching_test.rb
@@ -312,18 +312,3 @@ class ViewCacheDependencyTest < ActionController::TestCase
     assert_equal %w(trombone flute), HasDependenciesController.new.view_cache_dependencies
   end
 end
-
-class DeprecatedPageCacheExtensionTest < ActiveSupport::TestCase
-  def test_page_cache_extension_binds_default_static_extension
-    deprecation_behavior = ActiveSupport::Deprecation.behavior
-    ActiveSupport::Deprecation.behavior = :silence
-    old_extension = ActionController::Base.default_static_extension
-
-    ActionController::Base.page_cache_extension = '.rss'
-
-    assert_equal '.rss', ActionController::Base.default_static_extension
-  ensure
-    ActiveSupport::Deprecation.behavior = deprecation_behavior
-    ActionController::Base.default_static_extension = old_extension
-  end
-end